Added an example custom preservation rule file
Added an example custom preservation rule file for the new
tool.htmlcompressor.opts.preserve option. The example rule file matches
mutiple lines and looks for head tags. In practice this may not be a
good thing to choose for preserving spacing, but it serves as an
example where the effect is easy to recongize.

Added a reference to the HtmlCompressor documentation with more details.